added chart and fixed linting errors
