+++ client/src/components/chartArea/ChartArea.js
@@ -6,11 +6,12 @@ import AllComparisonCharts from '../allComparisonCharts/AllComparisonCharts';
 import PropTypes from 'prop-types';
 
 const ChartArea = ({ data }) => {
-  
+
   const mobileData = data.filter(mobile => mobile.device === 'Mobile devices with full browsers');
   const mobileHours = mobileData.map(hour => hour.hourOfDay).reverse();
   const maxImpressions = mobileData.map(mobile => mobile.maxImpressions).reverse();
   const tabletData = data.filter(mobile => mobile.device === 'Tablets with full browsers');
+  const tabletHours = tabletData.map(hour => hour.hourOfDay).reverse();
   const cost = tabletData.map(cost => cost.cost).reverse();
   const searchImpressions = tabletData.map(searchImp => searchImp.searchImpressionShare);
 
@@ -23,7 +24,7 @@ const ChartArea = ({ data }) => {
         <AllMobileCharts data={data} hours={mobileHours} max={maxImpressions} />
       </div>
       <div className='all-tablet'>
-        <AllTabletCharts data={data} cost={cost} search={searchImpressions} />
+        <AllTabletCharts data={data} cost={cost} search={searchImpressions} hours={tabletHours} />
       </div>
       <div className='all-comparison'>
         <AllComparisonCharts data={data} />
